fix(o2k): use regex_priority in defaults if given
This default should be left untouched unless you know what you are doing
(same as strip_path). But the value will be honored now if given.
The default is assumed to be the plain-path value. A parameterized
path should have a lower precedence, hence if the value is given
a parameterized path will get the priority-1.