fix(o2k): limit path-param names to 32 chars (#153)
PCRE engines have a limit of 32 characters.

the path-parameter-names need to be predictable, because they might reuse them in other
places (request-transformer template for example). So truncation or truncation and replacing
the last x-chars with a hash for uniqueness, will not work, since the names become unpredictable.

Then I think the only alternative is to throw an error in OAS2kong, and have the user preprocess
them to shorter names, such that they are well informed about the new names in use. More work,
but no surprises.